+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <a name="v3.0.0"></a>
  # v3.0.0
+### Feature
+* Replaces worker processes that crash/exit unexpectedly
 ### Security
 * Updated dependencies to newer versions
 ### Removed